server/clip-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
torch>=2.0.0
torchvision>=0.15.0
clip @ git+https://github.com/openai/CLIP.git
ftfy>=6.0.0
regex>=2023.0.0
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.24.0
